pyproject.toml
[project]
name = "task-3-currency-exchange"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "fastapi[standard]>=0.120.0",
    "pillow>=12.0.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.44",
]
